consumer/requirements.txt
#pymorphy2-dicts-uk==2.4.1.1.1460299261
#pyspark==2.4.5
kafka-python==2.0.1
